+++ main/joystick.h
@@ -1,9 +1,11 @@
 #ifndef __JOYSTICK_H__
 #define __JOYSTICK_H__
 
+#include <stdint.h>
+
 typedef struct {
-    int x_percent;
-    int y_percent;
+    int8_t x_percent;
+    int8_t y_percent;
     int power_switch;
 } Joystick;
